requirements.txt
transformers==4.38.2
torch==2.2.1
pandas==2.2.1
numpy==1.26.4
pillow==10.2.0
